+++ R/merge_requests.R
@@ -17,7 +17,7 @@
 #' # Create MR and get its information
 #' mr_infos <- gl_create_merge_request(project = <<your-project-id>>, 
 #'   source_branch = "my-extra-branch",
-#'   title = "Merge extra to master", description = "These modifications are wonderful")
+#'   title = "Merge extra to main", description = "These modifications are wonderful")
 #' # List all opened MR
 #' gl_list_merge_requests(project = <<your-project-id>>, status = "opened")
 #' # Edit MR created
@@ -28,7 +28,7 @@
 #' # Delete MR as it never existed
 #' gl_delete_merge_request(project = <<your-project-id>>, merge_request_iid = mr_infos$iid)
 #' }
-gl_create_merge_request <- function(project, source_branch, target_branch = "master", title, description, ...) {
+gl_create_merge_request <- function(project, source_branch, target_branch = get_main(), title, description, ...) {
   gl_proj_req(project = project, c("merge_requests"), ...) %>% 
     gitlab(source_branch = source_branch,
            target_branch = target_branch,